split content into open and manage
