+++ features/omni-kit/protocols/erc-4626/components/details-section/index.ts
@@ -2,4 +2,6 @@ export * from './Erc4626DetailsSectionContent'
 export * from './Erc4626DetailsSectionContentAllocation'
 export * from './Erc4626DetailsSectionContentEstimatedEarnings'
 export * from './Erc4626DetailsSectionContentEstimationLoader'
+export * from './Erc4626DetailsSectionContentManage'
+export * from './Erc4626DetailsSectionContentOpen'
 export * from './Erc4626DetailsSectionFooter'
